collapsed constants files, moved into root
